requirements.txt
# dependencies
colorama
openai
python-dotenv
uvicorn
pydantic
fastapi
python-multipart
langchain
langchain-openai
langchain_community
requests
SQLAlchemy
pytest
pytest-asyncio
langchain-community
langchain-experimental
langsmith
PyPDF2
langchain_qdrant
psycopg2
starlette
passlib[bcrypt]
pydantic[email]
langgraph
langgraph-checkpoint-postgres
IPython
asyncpg
psycopg